requirements.txt
slackclient
slackbot
requests